+++ src/widget/mod.rs
@@ -9,7 +9,7 @@ pub mod periodic;
 pub mod mpd;
 #[cfg(feature = "systemstat")] pub mod delayed;
 #[cfg(feature = "xkb")] pub mod xkb;
-#[cfg(feature = "unix_alsa")] pub mod alsa;
+pub mod volume;
 
 pub use self::base::*;
 pub use self::text::*;
@@ -18,6 +18,6 @@ pub use self::bspwm::*;
 pub use self::datetime::*;
 pub use self::periodic::*;
 pub use self::mpd::*;
+pub use self::volume::*;
 #[cfg(feature = "systemstat")] pub use self::delayed::*;
 #[cfg(feature = "xkb")] pub use self::xkb::*;
-#[cfg(feature = "unix_alsa")] pub use self::alsa::*;
